gyoumu-backend/requirements.txt
Django
djangorestframework
django-cors-headers
snowflake-connector-python
